+++ src/test/groovy/ru/vyarus/dropwizard/guice/test/LifecycleStartedForGuiceyTest.groovy
@@ -2,9 +2,11 @@ package ru.vyarus.dropwizard.guice.test
 
 import io.dropwizard.Application
 import io.dropwizard.Configuration
+import io.dropwizard.setup.Bootstrap
 import io.dropwizard.setup.Environment
 import org.eclipse.jetty.util.component.AbstractLifeCycle
 import org.eclipse.jetty.util.component.LifeCycle
+import ru.vyarus.dropwizard.guice.GuiceBundle
 import ru.vyarus.dropwizard.guice.test.spock.UseGuiceyApp
 import spock.lang.Specification
 
@@ -26,6 +28,11 @@ class LifecycleStartedForGuiceyTest extends Specification {
     }
 
     static class App extends Application<Configuration> {
+
+        @Override
+        void initialize(Bootstrap<Configuration> bootstrap) {
+            bootstrap.addBundle(GuiceBundle.builder().build())
+        }
 
         @Override
         void run(Configuration configuration, Environment environment) throws Exception {