+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Add junit 5 extensions (#74). Works much like existing spock extensions:
     - @TestGuiceyApp for replacement of GuiceyAppRule
     - @TestDropwizardApp for using instead of DropwizardAppRule (or current dropwizard extension)
+* Spock extensions internally use DropwizardTestSupport instead of deprecated junit 4 rules    
 * Fix parallel tests support: guice logs interception wasn't thread safe (#103)     
 
 ### 5.0.1 (2020-03-13)